-Added threading support to windows port. -Added portable socket code. -Reworked gdb stub so it uses portable threading code. -Reworked Netlink emulation. Now it's multi-threaded and should work beyond local and LAN connections. -Changed Netlink debug logging so it's separate. -A few CMakeLists.txt tweaks for SDL2 and winsock on windows.
git-svn-id: svn+ssh://svn.code.sf.net/p/yabause/code/trunk@3144 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/gdb/packet.c
@@ -3,7 +3,7 @@
 
 void gdb_packet_read(gdb_packet * packet, char * buffer, size_t got)
 {
-   ssize_t i;
+   size_t i;
    char c;
 
    for(i = 0;i < got;i++)
@@ -37,7 +37,7 @@ void gdb_packet_read(gdb_packet * packet, char * buffer, size_t got)
             packet->checksum[packet->cslen++] = c;
             if (2 == packet->cslen)
             {
-               u8 checksum;
+               int checksum;
                sscanf(packet->checksum, "%2x", &checksum);
                packet->buffer[packet->buflen] = '\0';
                if (packet->bufsum == checksum)
